requirements.txt
fastapi
pinecone
google-genai
numpy
python-dotenv
mangum